Move azle canister
