+++ examples/motoko_examples/factorial/canisters/azle/index.did
