+++ examples/motoko_examples/factorial/test/test.ts
@@ -1,5 +1,5 @@
 import { deploy, run_tests, Test } from 'azle/test';
-import { createActor } from '../test/dfx_generated/factorial';
+import { createActor } from '../test/dfx_generated/azle';
 
 const factorial_canister = createActor('rrkah-fqaaa-aaaaa-aaaaq-cai', {
     agentOptions: {
@@ -8,7 +8,7 @@ const factorial_canister = createActor('rrkah-fqaaa-aaaaa-aaaaq-cai', {
 });
 
 const tests: Test[] = [
-    ...deploy('factorial'),
+    ...deploy('azle'),
     {
         name: '0 factorial',
         test: async () => {
